Separate util test out
