+++ packages/ember-routing/lib/location/auto_location.js
@@ -15,7 +15,6 @@ import {
   replacePath
 } from "ember-routing/location/util";
 
-
 /**
 @module ember
 @submodule ember-routing
